+++ components/frontend/src/subject/Subjects.js
@@ -12,6 +12,8 @@ export function Subjects(props) {
   const [hideMetricsNotRequiringAction, setHideMetricsNotRequiringAction] = useURLSearchQuery(props.history, "hide_metrics_not_requiring_action", "boolean");
   const [hiddenColumns, toggleHiddenColumn, clearHiddenColumns] = useURLSearchQuery(props.history, "hidden_columns", "array");
   const [visibleDetailsTabs, toggleVisibleDetailsTab, clearVisibleDetailsTabs] = useURLSearchQuery(props.history, "tabs", "array");
+  const [trendTableNrDates, setTrendTableNrDates] = useURLSearchQuery(props.history, "trend_table_nr_dates", "integer", 7);
+  const [trendTableInterval, setTrendTableInterval] = useURLSearchQuery(props.history, "trend_table_interval", "integer", 1);
   const last_index = Object.keys(props.report.subjects).length - 1;
   return (
     <>
@@ -27,9 +29,13 @@ export function Subjects(props) {
             key={subject_uuid}
             last_subject={index === last_index}
             setHideMetricsNotRequiringAction={(state) => setHideMetricsNotRequiringAction(state)}
+            setTrendTableNrDates={(nr) => setTrendTableNrDates(nr)}
+            setTrendTableInterval={(interval) => setTrendTableInterval(interval)}
             subject_uuid={subject_uuid}
             toggleHiddenColumn={(column) => toggleHiddenColumn(column)}
             toggleVisibleDetailsTab={(...tabs) => toggleVisibleDetailsTab(...tabs)}
+            trendTableInterval={trendTableInterval}
+            trendTableNrDates={trendTableNrDates}
             visibleDetailsTabs={visibleDetailsTabs}
           /> : null
       )}